+++ src/md_xtc.c
@@ -255,7 +255,7 @@ static size_t xtc_fetch_frame_data(struct md_trajectory_o* inst, int64_t frame_i
         return 0;
     }
 
-    if (frame_idx < 0 || xtc->header.num_frames <= frame_idx) {
+    if (frame_idx < 0 || (int64_t)xtc->header.num_frames <= frame_idx) {
         MD_LOG_ERROR("XTC: Frame index is out of range");
         return 0;
     }
